feat(flags): add flag to force never response body
The OpenAPI specs state that in order to describe an Empty Response Body, the response `content` section should be omitted.
https://swagger.io/docs/specification/describing-responses/

The current behavior generates the expected `never` type for a handful of status code which should always return an empty response body.
However, when an app returns an empty response body for a status code outside of this short list, we lose type safety.

This flag allows developers to generate types with the `never` type used for every response with no `content`